~unhardcoding of user secrets ...

+++ Src/CI/GenderApiLib/GenderApi.cs
@@ -1,14 +1,10 @@
-﻿using System.Diagnostics;
-using System.Security.Policy;
-using GenderApiLib.Model;
-using Microsoft.Extensions.Configuration;
-using Newtonsoft.Json;
-
-namespace GenderApiLib;
+﻿namespace GenderApiLib;
 public class GenderApi
 {
-  public static async Task<(TimeSpan ts, string finishReason, FirstnameRootObject? root)> CallOpenAI(IConfigurationRoot cfg, string firstName)
+  public static async Task<(TimeSpan ts, string exMsg, FirstnameRootObject? root)> CallOpenAI(IConfigurationRoot cfg, string firstName)
   {
+    Trace.WriteLine($"■ ■ ■ cfg?[\"WhereAmI\"]: '{cfg?["WhereAmI"]}'.");
+
     var stopwatch = Stopwatch.StartNew();
     var filename = $@"C:\g\CI\Src\CI\GenderApiLib\Cache.FirstName\{firstName}.json";
 
@@ -18,14 +14,17 @@ public class GenderApi
       if (jsonPart is null)
         return (stopwatch.Elapsed, "jsonPart is null", null);
 
-      if (!File.Exists(filename))
-        await File.WriteAllTextAsync(filename, jsonPart);
+      //if (jsonPart.Contains("errno")) return (stopwatch.Elapsed, "jsonPart has errno", null);
+      //if (jsonPart.Contains("errmsg")) return (stopwatch.Elapsed, "jsonPart has errmsg", null);
 
-      var dynOb2 = JsonConvert.DeserializeObject<FirstnameRootObject>(jsonPart);          //dynamic dynObj = JsonConvert.DeserializeObject(jsonPart) ?? "No way!";          //dynamic dynObj = JsonSerializer.Deserialize(json);
+      var root = JsonConvert.DeserializeObject<FirstnameRootObject>(jsonPart);          //dynamic dynObj = JsonConvert.DeserializeObject(jsonPart) ?? "No way!";          //dynamic dynObj = JsonSerializer.Deserialize(json);
+
+      if (string.IsNullOrEmpty(root?.errmsg) && !File.Exists(filename) )
+        await File.WriteAllTextAsync(filename, jsonPart);
 
       try
       {
-        return (stopwatch.Elapsed, "", dynOb2);
+        return (stopwatch.Elapsed, "", root);
       }
       catch (Exception ex)
       {
